addSignature button moved to the insideHearing component
