+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/hearings/src/pages/employee/index.js
@@ -21,7 +21,6 @@ import ViewPendingTask from "./ViewPendingTask";
 import HearingPopup from "./HearingPopUp";
 import InsideHearing from "./InsideHearing";
 import ViewCase from "./ViewCase";
-import HelloPage from "./HelloPage"
 const bredCrumbStyle={ maxWidth: "min-content" };
 const ProjectBreadCrumb = ({ location }) => {
   const { t } = useTranslation();
@@ -66,7 +65,6 @@ const App = ({ path, stateCode, userType, tenants }) => {
         <PrivateRoute path={`${path}/view-witness-deposition`} component={() => <ViewWitnessDeposition></ViewWitnessDeposition>} />
         <PrivateRoute path={`${path}/view-pending-task`} component={() => <ViewPendingTask></ViewPendingTask>} />
         <PrivateRoute path={`${path}/submission`} component={() => <Submission></Submission>} />
-        <PrivateRoute path={`${path}/say-hello`} component={() => <HelloPage></HelloPage>} />
       </AppContainer>
     </Switch>
   );
